==> 7_differential_expression.Rmd <==
---
title: "differential_expression"
output: html_document
date: "2023-09-18"
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "~/Alyssa/Mouse_Kidney/")

library(tidyverse)
library(Seurat)
library(DESeq2)
library(SingleCellExperiment)
library(scuttle)
library(apeglm)
```

# Get counts.
```{r}
renamed_seu = readRDS("rds_objects/7_renamed_seu_sc.rds")
counts = renamed_seu@assays$RNA
```

# Functions for differential expression.
```{r}
clump = function(agg_sce, label1, label2, cell_type) {
    agg_sce = agg_sce[, agg_sce$Condition %in% c(label1, label2)] # Subset on group of interest.

    metadata = colData(agg_sce)[,c("Condition", "metacell")]
    metadata$CHIP = factor(metadata$Condition, levels = c(label2, label1)) # Establish to get the order we want.
    
    dds = DESeqDataSetFromMatrix(round(assay(agg_sce, "counts")), 
                    colData = metadata, 
                    design = ~ Condition)
    
    keep = rowSums(counts(dds)) >= 3
    dds = dds[keep,]
    dds = DESeq(dds)
    
    res = results(dds, contrast = c("Condition", label1, label2))

    summary(res)
    res_ordered_wald = res[order(res$padj),] %>%
      as.data.frame() %>%
      filter(!is.na(padj)) %>%
      rownames_to_column("gene")
    write_tsv(res_ordered_wald, paste0("tables/metacells_de/", cell_type, "_", label1, "_vs_", label2, ".tsv"))
    return(res_ordered_wald)
}


get_cell_type_data_for_clumping = function(cell_type_name) {
  ko_cell_type_metadata = read_csv(paste0("metacells_output/ko_", gsub(' ',  '_', cell_type_name), "_metacells.csv")) %>%
    mutate(Condition = "KO")
  wt_cell_type_metadata = read_csv(paste0("metacells_output/wt_", gsub(' ',  '_', cell_type_name), "_metacells.csv")) %>%
    mutate(Condition = "WT")
  cell_type_metadata = rbind(ko_cell_type_metadata, wt_cell_type_metadata) %>%
    dplyr::rename(Cell_ID = `...1`) %>%
    filter(metacell >= 0) # negative metacell assignments given to outliers

  cell_type_counts = counts@counts[,cell_type_metadata$Cell_ID] # this syntax changes for seurat v4 and seurat v5 (counts vs counts@counts)
  rownames(cell_type_metadata) = cell_type_metadata$Cell_ID

  cell_type_sce = SingleCellExperiment(cell_type_counts, colData = cell_type_metadata)
  assayNames(cell_type_sce) = "counts"

  cell_type_agg_sce = aggregateAcrossCells(cell_type_sce, ids = colData(cell_type_sce)[,c("Condition", "metacell")])
  return(cell_type_agg_sce)
}


```


# Run differential expression.
```{r}
run_chip_comparisons = function(cell_type_name) {

  agg_cell_type_data = get_cell_type_data_for_clumping(cell_type_name)

  tryCatch({
    chip_comparison = clump(agg_cell_type_data, label1 = "KO", label2 = "WT", cell_type = cell_type_name) 
  }, error = function(cond) {
    message("CHIP comparison failed.")
  })
}

cell_types = renamed_seu$predicted.id %>% unique()
lapply(cell_types, run_chip_comparisons)
run_chip_comparisons("t cell")
```

```{r}
run_chip_comparisons("neutrophil")
```
